+++ packages/nocodb-sdk/src/lib/index.ts
@@ -8,6 +8,7 @@ export * from '~/lib/helperFunctions';
 export * from '~/lib/enums';
 export * from '~/lib/formulaHelpers';
 export * from '~/lib/regex';
+export * from '~/lib/payment';
 export {
   default as UITypes,
   UITypesName,
